# docs/swagger.yaml
basePath: /
definitions:
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
  model.User:
    properties:
      address:
        type: string
      avatarUrl:
        type: string
      createdAt:
        type: string
      deletedAt:
        $ref: '#/definitions/gorm.DeletedAt'
      email:
        type: string
      id:
        type: integer
      nickname:
        type: string
      password:
        type: string
      personId:
        type: integer
      phone:
        type: string
      role:
        default: ROLE_USER
        type: string
      updatedAt:
        type: string
      username:
        type: string
    required:
    - username
    type: object
  model.UserLoginTemp:
    properties:
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
  model.UserRegisterTemp:
    properties:
      nickname:
        type: string
      password:
        type: string
      username:
        type: string
    required:
    - password
    - username
    type: object
host: localhost:9090
info:
  contact:
    email: grezkl@protonmail.com
    name: grezkl
    url: http://github.com/grezkl
  description: e-gov server
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: e-gov server
  version: "1.0"
paths:
  /affairPage:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页数
        in: query
        name: pageNum
        required: true
        type: integer
      - description: 页大小
        in: query
        name: pageSize
        required: true
        type: integer
      - description: 政务名称
        in: query
        name: name
        type: string
      responses: {}
      summary: 政务管理分页
      tags:
      - Affair
...
  /delUserBatch:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户编号数组
        in: path
        items:
          type: integer
        name: ids
        required: true
        type: array
      responses: {}
      summary: 注销用户（批量）
      tags:
      - Account
    get:
      consumes:
      - application/json
      parameters:
      - description: 角色
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据角色查询用户
      tags:
      - Account
...
  /deleteUserById/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 注销用户（单个）
      tags:
      - Account
...
  /login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserLoginTemp'
      responses: {}
      summary: 账号登录
      tags:
      - Account
...
  /pullUserById/{id}:
    get:
      consumes:
      - application/json
      parameters:
      - description: 政务申请编号
        in: path
        name: id
        required: true
        type: integer
      responses: {}
      summary: 根据编号查找政务申请
      tags:
      - Account
...
  /register:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号信息
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/model.UserRegisterTemp'
      responses: {}
      summary: 账号注册
      tags:
      - Account
...
  /saveUser:
    post:
      consumes:
      - application/json
      parameters:
      - description: 保存账号信息
        in: body
        name: user_form
        required: true
        schema:
          $ref: '#/definitions/model.User'
      responses: {}
      summary: 保存账号信息
      tags:
      - Account
...
  /updatePWD:
    post:
      consumes:
      - application/json
      parameters:
      - description: 账号
        in: body
        name: username
        required: true
        schema:
          type: string
      - description: 密码
        in: body
        name: password
        required: true
        schema:
          type: string
      responses: {}
      summary: 修改密码
      tags:
      - Account
...
  /userImport:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: 文件
        in: formData
        name: file
        required: true
        type: file
      responses: {}
      summary: 导入用户数据
      tags:
      - Account
  /username/:name:
    get:
      consumes:
      - application/json
      parameters:
      - description: 账号
        in: path
        name: username
        required: true
        type: string
      responses: {}
      summary: 获取用户信息
      tags:
      - Account
swagger: "2.0"